last commit

+++ app/src/main/java/com/nick/sampleffmpeg/Define/Constant.java
@@ -15,7 +15,7 @@ public class Constant {
     public final static int			BUTTON_NORMAL_ALPHA	= 100;
     public final static int			BUTTON_FOCUS_ALPHA	= 50;
     public static int               TIMELINE_UNIT_SECOND = 1;
-    public static float               SP_PER_SECOND = 60.f / TIMELINE_UNIT_SECOND;
+    public static int               SP_PER_SECOND = 60 / TIMELINE_UNIT_SECOND;
     public final static String      APP_NAME = "VideoEditorApp";
     public static String strTemporaryFolderName = Long.toString(System.currentTimeMillis());
 
@@ -33,7 +33,7 @@ public class Constant {
 
     public static void updateTimeUnit(int time) {
         TIMELINE_UNIT_SECOND = time;
-        SP_PER_SECOND = 60.0f / TIMELINE_UNIT_SECOND;
+        SP_PER_SECOND = 60 / TIMELINE_UNIT_SECOND;
     }
 
     public static void setVideoSize(int width, int height) {
